projectのschemaをdomainディレクトリ以下に移動
